# Build multiple binaries in single command