This speeds up builds 5-10x

Timing:

Cached: 6s -> 1s
Clean: 40s -> 7s

[x] Test and Release

Marking as WIP because this probably breaks tests -- needs a lot more thought