@istio-testing 2019-09-17
@howardjohn: The following test **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lintmodern_istio | a1a9cee1b0a3b5c91e1416fec254e117d130765b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/17153/lintmodern_istio/197) | `/test lintmodern`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->