@istio-testing 2019-09-19
@howardjohn: The following tests **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-security-k8s-tests_istio | a1a9cee1b0a3b5c91e1416fec254e117d130765b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/17153/integ-security-k8s-tests_istio/468) | `/test integ-security-k8s-tests`
integ-new-install-k8s-tests_istio | 98d1e7b6cdb87f19a0c51a53f9a5847508692a9b | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/17153/integ-new-install-k8s-tests_istio/376) | `/test integ-new-install-k8s-tests`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->