@howardjohn 2019-09-20
@istio/wg-policies-and-telemetry-maintainers mind taking a look?